+++ src/features/scoring/index.js
@@ -8,6 +8,7 @@ export { default as StandingsSelfRecapCard } from './ui/StandingsSelfRecapCard';
 export { default as StandingsOfficialSetlistCard } from './ui/StandingsOfficialSetlistCard';
 export { computeStandingsSelfRecap } from './model/standingsSelfRecap';
 export { groupOfficialSetlistBySet } from './model/groupOfficialSetlistBySet';
+export { buildPhishNetSetlistUrl } from './model/buildPhishNetSetlistUrl';
 export { computeShowWinnerOfTheNight } from './model/useShowWinnerOfTheNight';
 export {
   GRADED_PICKS_SHARE_DOMAIN,
@@ -38,6 +39,7 @@ export { default as StandingsPoolPicker } from './ui/StandingsPoolPicker';
 export { default as StandingsShowOrPoolView } from './ui/StandingsShowOrPoolView';
 export { default as StandingsTourView } from './ui/StandingsTourView';
 export { default as StandingsViewToggle } from './ui/StandingsViewToggle';
+export { default as StandingsStickyChrome } from './ui/StandingsStickyChrome';
 export { default as StandingsWinnerOfTheNightBanner } from './ui/StandingsWinnerOfTheNightBanner';
 export { default as TourStandingsSection } from './ui/TourStandingsSection';
 export { resolveCurrentTour } from './model/resolveCurrentTour';